First Task added
Write a Script to create 100 folders within the given directory.
The name of each folder should be folder1/ folder2/ and so on till folder100/
Inside each folderX/ there should be an empty file named folderX.txt (eg. folder1.txt)

It should be such that ONLY THE OWNER of the files and folders should have read-write-executable permissions.